+++ src/lib.rs
@@ -16,6 +16,7 @@ pub mod once;
 pub mod proc;
 pub(crate) mod psi;
 pub mod root;
+pub(crate) mod sixel;
 pub(crate) mod tty;
 pub mod types;
 pub mod ui;